refactor: rename app package
