+++ README.md
@@ -7,4 +7,4 @@ Settings are read from environment variables. Notable options include:
 - `REMINDER_INTERVAL_SECONDS`: Interval (in seconds) between reminder checks for
   upcoming tasks. Defaults to `60`.
 
-See `app/core/config.py` for a full list of configurable options.
+See `task_manager_app/core/config.py` for a full list of configurable options.
