--- app/models/__init__.py
@@ -1,3 +1,0 @@
-from app.models.user import User, Role
-from app.models.task import Task, Priority, Status
-from app.models.notification import Notification
