+++ task_manager_app/core/database.py
@@ -1,6 +1,6 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, declarative_base
-from app.core.config import get_settings
+from task_manager_app.core.config import get_settings
 
 settings = get_settings()
 db_url = settings.database_url  # <- lowercase property
