+++ task_manager_app/models/__init__.py
@@ -1,0 +1,3 @@
+from task_manager_app.models.user import User, Role
+from task_manager_app.models.task import Task, Priority, Status
+from task_manager_app.models.notification import Notification
