+++ task_manager_app/routers/auth.py
@@ -3,8 +3,8 @@ from fastapi.security import OAuth2PasswordRequestForm
 from jose import JWTError
 from sqlalchemy import select
 from sqlalchemy.orm import Session
-from app.core.database import get_db
-from app.core.security import (
+from task_manager_app.core.database import get_db
+from task_manager_app.core.security import (
     hash_password,
     verify_password,
     create_access_token,
@@ -12,8 +12,8 @@ from app.core.security import (
     verify_refresh_token,
     revoke_refresh_token,
 )
-from app.models.user import User, Role
-from app.schemas.user import UserCreate, Token, TokenRefresh
+from task_manager_app.models.user import User, Role
+from task_manager_app.schemas.user import UserCreate, Token, TokenRefresh
 
 router = APIRouter(prefix="/auth", tags=["auth"])
