+++ task_manager_app/routers/tasks.py
@@ -3,11 +3,11 @@ from typing import Optional, List, Annotated
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.orm import Session
 from sqlalchemy import select, and_, or_
-from app.core.database import get_db
-from app.models.task import Task, Priority, Status
-from app.models.user import User, Role
-from app.schemas.task import TaskCreate, TaskUpdate, TaskOut
-from app.routers.users import get_current_user
+from task_manager_app.core.database import get_db
+from task_manager_app.models.task import Task, Priority, Status
+from task_manager_app.models.user import User, Role
+from task_manager_app.schemas.task import TaskCreate, TaskUpdate, TaskOut
+from task_manager_app.routers.users import get_current_user
 
 router = APIRouter(prefix="/tasks", tags=["tasks"])
